Move the source around to allow easier sharing.
