+++ LottieGen/LottieGen.sln
@@ -107,14 +107,18 @@ Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Animatables", "..\source\An
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LottieGen.win-x64", "win-x64\LottieGen.win-x64.csproj", "{FD20C604-C757-46F4-9A8C-711C39F79A46}"
 EndProject
-Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "LottieGen.Common", "Common\LottieGen.Common.shproj", "{54FB5765-962C-432C-B001-56D20EBE8C8E}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LottieGen", "DotnetTool\LottieGen.csproj", "{A435DD2A-942F-4D95-9E3F-216FFA5FBC4D}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{83310E27-6098-4DC2-8EFE-EC7D9418E2E5}"
 	ProjectSection(SolutionItems) = preProject
 		README.md = README.md
 	EndProjectSection
+EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "LottieGen", "..\source\LottieGen\LottieGen.shproj", "{7654A857-9A99-4185-9F8E-DD0CE662AF23}"
+EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "CommandLine", "..\source\CommandLine\CommandLine.shproj", "{DB8248EB-8DD0-4EE0-86E3-F0BA9A405933}"
+EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "LottieGenExe", "..\source\LottieGenExe\LottieGenExe.shproj", "{DC7F7A39-B7A7-40B4-937A-601460CAE38C}"
 EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
@@ -124,15 +128,51 @@ Global
 		..\source\WinUIXamlMediaData\WinUIXamlMediaData.projitems*{30059ca7-0745-4eec-8d11-b14850a70c98}*SharedItemsImports = 13
 		..\source\YamlData\YamlData.projitems*{39c6b7f3-5e75-4019-82ab-00fd8a0a06e2}*SharedItemsImports = 13
 		..\source\LottieReader\LottieReader.projitems*{4e7d8957-3f5f-46e1-99a8-2012b806c9b0}*SharedItemsImports = 13
-		Common\LottieGen.Common.projitems*{54fb5765-962c-432c-b001-56d20ebe8c8e}*SharedItemsImports = 13
 		..\source\DotLottie\DotLottie.projitems*{7012420d-624c-4bd4-a1d2-1c6c1655ed3a}*SharedItemsImports = 13
 		..\source\UIData\UIData.projitems*{74601e6c-2dfe-4842-b170-047941abff2c}*SharedItemsImports = 13
+		..\source\LottieGen\LottieGen.projitems*{7654a857-9a99-4185-9f8e-dd0ce662af23}*SharedItemsImports = 13
 		..\source\GenericData\GenericData.projitems*{77bcd724-8555-463b-985f-f8e8110164c4}*SharedItemsImports = 13
+		..\source\Animatables\Animatables.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\CommandLine\CommandLine.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\CompMetadata\CompMetadata.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\DotLottie\DotLottie.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\GenericData\GenericData.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\LottieData\LottieData.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\LottieGenExe\LottieGenExe.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\LottieGen\LottieGen.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\LottieMetadata\LottieMetadata.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\LottieReader\LottieReader.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\LottieToWinComp\LottieToWinComp.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\UIDataCodeGen\UIDataCodeGen.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\UIData\UIData.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\WinCompData\WinCompData.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\WinStorageStreamsData\WinStorageStreamsData.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\WinUIXamlMediaData\WinUIXamlMediaData.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
+		..\source\YamlData\YamlData.projitems*{a435dd2a-942f-4d95-9e3f-216ffa5fbc4d}*SharedItemsImports = 5
 		..\source\CompMetadata\CompMetadata.projitems*{b0197c19-bdf5-473e-a022-e21f6122eee5}*SharedItemsImports = 13
 		..\source\LottieData\LottieData.projitems*{b3db16ee-a821-4474-a188-e64926529bbd}*SharedItemsImports = 13
 		..\source\WinCompData\WinCompData.projitems*{d02870de-7ded-4916-85d4-3175ceedef74}*SharedItemsImports = 13
 		..\source\UIDataCodeGen\UIDataCodeGen.projitems*{d02be6c8-14db-4b4f-8600-f3c9b69c104d}*SharedItemsImports = 13
+		..\source\CommandLine\CommandLine.projitems*{db8248eb-8dd0-4ee0-86e3-f0ba9a405933}*SharedItemsImports = 13
+		..\source\LottieGenExe\LottieGenExe.projitems*{dc7f7a39-b7a7-40b4-937a-601460cae38c}*SharedItemsImports = 13
 		..\source\Animatables\Animatables.projitems*{fc89273a-b2da-4625-8a73-ef02a658d65e}*SharedItemsImports = 13
+		..\source\Animatables\Animatables.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\CommandLine\CommandLine.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\CompMetadata\CompMetadata.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\DotLottie\DotLottie.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\GenericData\GenericData.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\LottieData\LottieData.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\LottieGenExe\LottieGenExe.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\LottieGen\LottieGen.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\LottieMetadata\LottieMetadata.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\LottieReader\LottieReader.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\LottieToWinComp\LottieToWinComp.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\UIDataCodeGen\UIDataCodeGen.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\UIData\UIData.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\WinCompData\WinCompData.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\WinStorageStreamsData\WinStorageStreamsData.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\WinUIXamlMediaData\WinUIXamlMediaData.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
+		..\source\YamlData\YamlData.projitems*{fd20c604-c757-46f4-9a8c-711c39f79a46}*SharedItemsImports = 5
 	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -207,6 +247,9 @@ Global
 		{B0197C19-BDF5-473E-A022-E21F6122EEE5} = {AB232F35-AAF7-4AE2-B1D2-45DD9BC2F7D7}
 		{7012420D-624C-4BD4-A1D2-1C6C1655ED3A} = {AB232F35-AAF7-4AE2-B1D2-45DD9BC2F7D7}
 		{FC89273A-B2DA-4625-8A73-EF02A658D65E} = {AB232F35-AAF7-4AE2-B1D2-45DD9BC2F7D7}
+		{7654A857-9A99-4185-9F8E-DD0CE662AF23} = {AB232F35-AAF7-4AE2-B1D2-45DD9BC2F7D7}
+		{DB8248EB-8DD0-4EE0-86E3-F0BA9A405933} = {AB232F35-AAF7-4AE2-B1D2-45DD9BC2F7D7}
+		{DC7F7A39-B7A7-40B4-937A-601460CAE38C} = {AB232F35-AAF7-4AE2-B1D2-45DD9BC2F7D7}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {51B9BB4C-5196-41CF-950C-12B04AD8A61C}
