+++ LottieGen/win-x64/LottieGen.win-x64.csproj
@@ -1,5 +1,4 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-  <Import Project="..\Common\CommonImports.proj" />
 
   <PropertyGroup>
     <PackageId>LottieGen.win-x64</PackageId>
@@ -19,10 +18,10 @@
   </PropertyGroup>
 
   <PropertyGroup>
-	<!--
-	  These properties cause the Publish target to convert the build output
-	  into a single .exe file that will run on Windows x64 with no dependencies.
-	-->
+  <!--
+    These properties cause the Publish target to convert the build output
+    into a single .exe file that will run on Windows x64 with no dependencies.
+  -->
     <SelfContained>true</SelfContained>
     <PublishSingleFile>true</PublishSingleFile>
     <PublishTrimmed>true</PublishTrimmed>
@@ -50,20 +49,16 @@
         <!-- Set IsPackable false so that Publish doesn't depend on Pack. -->
         <IsPackable>false</IsPackable>
       </PropertyGroup>
-      <CallTarget Targets="Publish"/>
+      <CallTarget Targets="Publish" />
   </Target>
 
   <!-- Generates the .nuspec file. -->
-  <Target
-    Name="LottieGenGenerateNuspec" 
-    Outputs="$(IntermediateOutputPath)$(PackageId).nuspec">
+  <Target Name="LottieGenGenerateNuspec" Outputs="$(IntermediateOutputPath)$(PackageId).nuspec">
     <PropertyGroup>
-       <NuspecFile>$(IntermediateOutputPath)$(PackageId).nuspec</NuspecFile>
-		<_LottieGenNuspecPropsFile>$(PackageId).props</_LottieGenNuspecPropsFile>
-    </PropertyGroup>
-    <ItemGroup>
-      <_NuSpecContent Include="_">
-           <Text><![CDATA[<?xml version="1.0" encoding="utf-8"?>
+      <NuspecFile>$(IntermediateOutputPath)$(PackageId).nuspec</NuspecFile>
+      <_LottieGenNuspecPropsFile>$(PackageId).props</_LottieGenNuspecPropsFile>
+      <_NuSpecContent>
+<![CDATA[<?xml version="1.0" encoding="utf-8"?>
 <package xmlns="http://schemas.microsoft.com/packaging/2012/06/nuspec.xsd">
   <metadata>
     <id>]]>$(PackageId)<![CDATA[</id>
@@ -84,9 +79,32 @@
     <file src="]]>$(MsBuildThisFileDirectory)$(_LottieGenNuspecPropsFile)<![CDATA[" target="]]>build/win10/$(_LottieGenNuspecPropsFile)<![CDATA[" />
   </files>
 </package>
-]]></Text>
+]]>
       </_NuSpecContent>
-    </ItemGroup>
-    <WriteLinesToFile File="$(NuspecFile)" Lines="@(_NuSpecContent->'%(Text)')" Overwrite="true"/>
+    </PropertyGroup>
+    <WriteLinesToFile File="$(NuspecFile)" Lines="$(_NuSpecContent)" Overwrite="true" />
   </Target>
+
+  <PropertyGroup>
+    <_SourceDir>..\..\source\</_SourceDir>
+    <StartupObject>Microsoft.Toolkit.Uwp.UI.Lottie.LottieGenExe.Program</StartupObject>
+  </PropertyGroup>
+  <Import Project="..\Common.proj" />
+  <Import Project="$(_SourceDir)Animatables\Animatables.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)CommandLine\CommandLine.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)CompMetadata\CompMetadata.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)DotLottie\DotLottie.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)GenericData\GenericData.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)LottieData\LottieData.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)LottieGen\LottieGen.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)LottieGenExe\LottieGenExe.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)LottieMetadata\LottieMetadata.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)LottieReader\LottieReader.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)LottieToWinComp\LottieToWinComp.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)UIData\UIData.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)WinCompData\WinCompData.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)WinStorageStreamsData\WinStorageStreamsData.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)WinUIXamlMediaData\WinUIXamlMediaData.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)YamlData\YamlData.projitems" Label="Shared" />
+  <Import Project="$(_SourceDir)UIDataCodeGen\UIDataCodeGen.projitems" Label="Shared" />
 </Project>
