+++ LottieGen/win-x64/Program.cs
@@ -1,0 +1,9 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+static class Program
+{
+    static int Main(string[] args)
+        => Microsoft.Toolkit.Uwp.UI.Lottie.LottieGenExe.Main.Run(args);
+}
