+++ source/LottieGen/Language.cs
@@ -1,0 +1,22 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Microsoft.Toolkit.Uwp.UI.Lottie.LottieGen
+{
+    enum Language
+    {
+        // Language wasn't recognized.
+        Unknown,
+
+        // Language specified was ambigious.
+        Ambiguous,
+
+        CSharp,
+        Cx,
+        Cppwinrt,
+        LottieYaml,
+        WinCompDgml,
+        Stats,
+    }
+}
